-Reorganized Package structure. Pulled out any classes or interfaces associated with Room and put them in a new "room" package
